Add support for ignoring PRs touching only non-subsantive files
See https://github.com/w3c/ash-nazg/issues/125#issuecomment-599390829

In the process, refactor in depth the spaghetti code used for PR validation:
* split in a separate module
* using await/async for readability (relying on doAsync to convert callback functions)

This means the code now uses a mix of callbacks and promises async code,
which isn't ideal; also, the GH module is partially able to send back promises -
a good first clean-up would be to make that code at least fully promise-able.

But in the spirit of incremental improvement, this is I think overall an improvement on the current code, making it easier to update the PR validation steps.