Limit who can update a stored token for a repo
The current code stores the user token for any repo they import, but that token is associated with the whole org rather than the single repo. This means that a user with admin on the repo but not the org will overrides the token and thus breaks the github status report on pull requests.

This change requires admin rights before updating the token, and rejects importing repos for orgs for which no admin token has been registered